Resolved issues in the owner view
