+++ wallet/lib/pages/owner_purchase_view_common/progress_bar_state.dart
@@ -1,0 +1,11 @@
+class ProgressBarState {
+  ProgressBarState({
+    required this.current,
+    required this.buffered,
+    required this.total,
+  });
+
+  final Duration current;
+  final Duration buffered;
+  final Duration total;
+}
