+++ wallet/lib/pylons_app.dart
@@ -17,7 +17,6 @@ import 'package:pylons_wallet/components/no_internet.dart';
 import 'package:pylons_wallet/components/pylons_app_theme.dart';
 import 'package:pylons_wallet/model/nft.dart';
 import 'package:pylons_wallet/pages/detailed_asset_view/owner_view.dart';
-import 'package:pylons_wallet/pages/detailed_asset_view/owner_view_view_model.dart';
 import 'package:pylons_wallet/pages/detailed_asset_view/widgets/pdf_viewer_full_screen.dart';
 import 'package:pylons_wallet/pages/home/home.dart';
 import 'package:pylons_wallet/pages/home/home_provider.dart';
@@ -126,10 +125,11 @@ class _PylonsAppState extends State<PylonsApp> with WidgetsBindingObserver {
 
                   if (ModalRoute.of(context)?.settings.arguments is NFT) {
                     final nft = ModalRoute.of(context)!.settings.arguments! as NFT;
-                    final viewModel = sl<OwnerViewViewModel>();
-                    viewModel.nft = nft;
+                    
+                    
                     return OwnerView(
-                      ownerViewViewModel: viewModel,
+                      key: ValueKey(nft),
+                      nft: nft,
                     );
                   }
